pyproject.toml
[project]
name = "aiconverse"
version = "0.1.0"
description = "Mini-Project AIConverse"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fpdf2>=2.8.3",
    "google-generativeai>=0.8.5",
    "pandas>=2.3.0",
    "pillow>=11.2.1",
    "plotly>=6.1.2",
    "python-dotenv>=1.1.1",
    "python-markdown>=0.1.0",
    "streamlit>=1.46.0",
]
